added the Helper class to get rid of the dependency between AI and entities moved TeamMode to the usecases
