+++ src/main/java/useCases/teammode/TeamBotPlayer.java
@@ -1,27 +1,24 @@
-package entities.teammode;
+package useCases.teammode;
 
-import entities.RealPlayer;
+import entities.BotPlayer;
 
-public class TeamRealPlayer extends RealPlayer implements TeamPlayer{
-
+public class TeamBotPlayer extends BotPlayer  implements TeamPlayer{
     private final Team teamOfPlayer;
-
     /**
      * creates a team player
-     *
      * @param name name
      * @param team team
      */
-    public TeamRealPlayer(String name, Team team) {
-        super(name);
+    public TeamBotPlayer(String name,String level, Team team){
+        super(name,level);
         teamOfPlayer = team;
         team.addTeamPlayer(this);
     }
     /**
      * returns team
-     * @return team
+     * @return the team
      */
-    public Team getTeam() {
+    public Team getTeam(){
         return this.teamOfPlayer;
     }
 }
