+++ src/main/java/useCases/teammode/TeamPlayer.java
@@ -1,4 +1,4 @@
-package entities.teammode;
+package useCases.teammode;
 
 public interface TeamPlayer {
     //to use the polymorphism of the clean architecture I created a teamMode interface which will help cast the player to this type
